Fixed a couple of bugs

+++ OakNet-Link-Api/PluginManager.cs
@@ -17,7 +17,7 @@ namespace OakNetLink.Api
                 var internalPlugin = new InternalPlugin();
                 plugins.Add(internalPlugin.pluginID, internalPlugin);
                 Logger.log("Registered internal plugin");
-                int id_ = internalPlugin.pluginID << 8;
+                ushort id_ = (ushort) (internalPlugin.pluginID << 8);
                 foreach (var kvp in internalPlugin.registerPackets())
                 {
                     PacketProcessor.addPacketProcessor(kvp.Value, kvp.Key, id_++);
@@ -26,7 +26,7 @@ namespace OakNetLink.Api
             if (plugins.ContainsKey(plugin.pluginID))
                 throw new Exception($"Plugin with ID: {plugin.pluginID} already registered!");
             plugins.Add(plugin.pluginID, plugin);
-            int id = plugin.pluginID << 8;
+            ushort id = (ushort) (plugin.pluginID << 8);
             Logger.log("Registered plugin with ID: " + plugin.pluginID);
             foreach (var kvp in plugin.registerPackets())
             {
